requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.4.6
aiohttp==3.11.13
aiohttp-retry==2.9.1
aiosignal==1.3.2
async-timeout==5.0.1
attrs==25.1.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
contourpy==1.3.0
cycler==0.12.1
flatbuffers==25.2.10
fonttools==4.56.0
frozenlist==1.5.0
idna==3.10
importlib_metadata==8.6.1
importlib_resources==6.5.2
jax==0.4.30
jaxlib==0.4.30
kiwisolver==1.4.7
matplotlib==3.9.4
mediapipe==0.10.21
ml_dtypes==0.5.1
multidict==6.1.0
numpy==1.26.4
opencv-contrib-python==4.11.0.86
opencv-python==4.11.0.86
opt_einsum==3.4.0
packaging==24.2
paho-mqtt==2.1.0
pillow==11.1.0
propcache==0.3.0
protobuf==4.25.6
pycparser==2.22
PyJWT==2.10.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
requests==2.32.3
scipy==1.13.1
sentencepiece==0.2.0
six==1.17.0
sounddevice==0.5.1
twilio==9.4.6
typing_extensions==4.12.2
urllib3==2.3.0
yarl==1.18.3
zipp==3.21.0
